+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - When using the Terraform deploy scripts as part of environment setup, workspace GUID values will no longer be wrapped in quotes
 - Storage Type is now required for all Power VS Instances and Volumes regardless of affinity
 - Dropdowns with only one possible item will now be automatically selected within a form
+- Currently, the `eu-es` Madrid region is not supported by Power Cloud API. Until support is added, Power VS Servers are disabled for the `eu-es` region on CRAIG
 
 ### Features
 
@@ -65,7 +66,6 @@ All notable changes to this project will be documented in this file.
 - Users can now import existing Secrets Manager instances into CRAIG
 - Power zones WDC06, WDC07, eu-de-1, and eu-de-2 now support Power Edge Router
 - Users can now create, update, and delete Fortigate VNF gateways from the `/form/fortigate` page
-- Power zones MAD02 (EU-ES-1), MAD04 (EU-ES-2), and SAO04 now support Power Edge Router
 
 ### Fixes